Organize generate_contacts backend
